+++ src/credential/index.ts
@@ -1,4 +1,4 @@
-import * as Issuing from "./issuing";
+import * as Issuance from "./issuance";
 import * as Presentation from "./presentation";
 
-export { Issuing, Presentation };
+export { Issuance, Presentation };
