+++ src/credential/issuance/04-complete-user-authorization.ts
