+++ src/credential/issuance/06-obtain-credential.ts
