+++ src/credential/presentation/02-evaluate-rp-trust.ts
@@ -1,6 +1,6 @@
 import { getRelyingPartyEntityConfiguration } from "../../trust";
 import { RelyingPartyEntityConfiguration } from "../../trust/types";
-import type { StartFlow } from "../issuing/01-start-flow";
+import type { StartFlow } from "../issuance/01-start-flow";
 import type { Out } from "../../utils/misc";
 
 export type EvaluateRelyingPartyTrust = (
